Add use case for creating merchants
The use case is considered here the unit, following once again the
approach suggested by Codely. That way, the repository or any other I/O
would be mocked/stubbed.

Therefore, the unit tests cover the creation of a merchant and every
single of its attributes are validated independently.

With the help of Dry gems every value object is validating its type and
sometimes its format.

These tests are currently in green.

References:

* https://dry-rb.org/
* https://dry-rb.org/gems/dry-types
* https://dry-rb.org/gems/dry-struct
* https://api.rubyonrails.org/v7.1.2/classes/ActiveSupport/Testing/TimeHelpers.html
* https://docs.rubocop.org/rubocop/cops_style.html#styleguardclause
* https://docs.rubocop.org/rubocop/cops_style.html#stylenegatedif
* https://docs.rubocop.org/rubocop/cops_metrics.html#metricsparameterlists
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecexamplelength
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecmessagespies
* https://docs.rubocop.org/rubocop-rspec/cops_rspec_rails.html#rspecrailstravelaround