Add logic to create order commissions
From now on, I'll add all the logic at once to be more agile. I've a
time constraint and have already shown with merchants and orders how to
do it incrementally with an outside-in TDD approach.

Ideally we'd publish a domain event when an order is created and we'd
have n event subscribers listening for that event. However, to keep it
as simple as possible here, a job is enqueued to create the
corresponding order commission.

I use the term commission to refer to the quantity that is charged to a
merchant after applying a fee. With access to a domain expert this kind
of naming choice would be easier.

The order commission stores the order amount, although it's not strictly
necessary to do it (denormalization).

I've defined different fee tiers for the commission that will be charged
to the merchant.

The commission is safely calculated using money-rails gem, that rounds
up to two decimals, as requested.

Once again, here I'm relying in the database constraints to ensure a
given order exists and no duplicated order commission is created for a
single order. In a real project probably we should ensure those constraints
at the application instead.

As acceptance test I'm reusing the acceptance test for orders (Rake
task) as a simpler solution. Using domain events we could use the event
subscriber instead.

Unit and integration tests are similar to the ones previously described
for merchants and orders.

References:

* https://www.quora.com/What-is-the-difference-between-commission-and-fees
* https://github.com/RubyMoney/money-rails
* https://github.com/sidekiq/sidekiq/wiki/Best-Practices
* https://stackoverflow.com/questions/45421201/ruby-json-library-convert-bigdecimal-to-scientific-notation