Add logic to create disbursements
Once again, here I'm relying in the database constraints to ensure a
given merchant exists. In a real project probably we should ensure
those constraints at the application level instead or in addition to it.

Add disbursement ID to orders, that is an optional attribute that will
be filled once a disbursement is created and will allow to know whether
the order has been disbursed or not.

As acceptance test I'm using the tests for the Rake task that will run
everyday.

This is a really complex scenario. I had to create lots of records to be
able to simulate how the creation of disbursements works and under what
circumnstances. Definitely there is room for improvement here. I tried
to help understand it a little bit with a few comments next to the
objects created.

Unit and integration tests are similar to the ones previously described
for the rest of resources.

References:

* https://api.rubyonrails.org/v7.1.3/classes/ActiveRecord/ConnectionAdapters/DatabaseStatements.html
* https://www.postgresql.org/docs/current/functions-aggregate.html
* https://www.postgresql.org/docs/current/functions-math.html
* https://stackoverflow.com/questions/34311457/how-does-rails-type-cast-the-result-of-array-agg-function
* https://api.rubyonrails.org/classes/ActiveSupport/Testing/TimeHelpers.html
* https://docs.rubocop.org/rubocop/cops_naming.html#namingvariablenumber
* https://dry-rb.org/gems/dry-types/1.7/constraints/
* https://dry-rb.org/gems/dry-types/1.7/optional-values/
* https://stackoverflow.com/questions/41054507/postgresql-array-of-elements-that-each-are-a-foreign-key