Allow to generate monthly fees for the first disbursement in the month for a merchant
That information is stored in separate table.

Here the use case for creating disbursements is used for generating
monthly fees when appropiate.

The logic to know if a monthly fee is applicable is placed in the
merchant entity, making that object a bit less anemic.

Due to current requirements, monthly fees are created after the first
disbursement of the month for a given merchant. That logic would need
to change for sure whenever the monthly fee needs to be substracted
from the disbursement amount.

Monthly fees are not created for merchants in the same month they go
live in the platform. The first one is created in their next month.

In the same way that happened with the update of orders' disbursement
ID, maybe a domain event could be used to generate these records, so
different modules could be less coupled.

The acceptance test for the Rake task that generates disbursements is
reused to check that orders are properly updated and monthly fees are
generated when appropiate.

It works because all those records are created or updated using a
background job, and the tag "sidekiq_inline" makes job to run
immediately in test environment.

I'd say that definitely not the best approach, but the easiest one in
this context. With more time, it's a part that probably could be
improved.

References:

* https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsdynamicfindby